Consistency fixes: pre-generated explosion audio, one threshold type
The explosion rumble was synthesized from scratch (a ~141 KB
allocation) on every event while pings were carefully pre-generated;
the rumble now gets one buffer at startup and playback clones an Arc,
matching the ping design. The identical noise across explosions is
imperceptible under the oscillator mix.

explosion_threshold was u64 in Config and usize in Simulation, bridged
by a cast; it is now usize end to end (RangedU64ValueParser keeps the
CLI range check).

Preset::label() hand-duplicates the clap-derived kebab-case names; a
test now pins the two together so a renamed variant cannot drift.

+++ src/sim.rs
@@ -205,9 +205,7 @@ impl Simulation {
             gravity_percent: config.gravity,
             wall_elasticity: config.wall_elasticity,
             particle_elasticity: config.particle_elasticity,
-            // Bounded by EXPLOSION_THRESHOLD_MAX (1000), which fits usize.
-            #[allow(clippy::cast_possible_truncation)]
-            explosion_threshold: config.explosion_threshold as usize,
+            explosion_threshold: config.explosion_threshold,
             matter: config.matter,
             flow: config.flow,
             particles: Vec::new(),
